Add additional miscellany to `about`.
This commit adds fields to the `AboutComponent` for various fun facts
and tidbits. May or may not convert to a more dynamic and future-proof
model with less hard-coding later, a la the skills component.

Fine-tuning of self-evaluation scaling for skills.

The `level` field of type `Skill` is now a defined `Level` type which
only accepts values between 1 and 5 to allow compile-time error-checking
if silly values are entered.